Create jounal-binding.h
